Split fixcall.rs into two parts for better APIs
